add token feature to tests POST and DELETE

+++ tests/testDB.test.js
@@ -2,20 +2,33 @@ const	supertest = require('supertest');
 const	app = require('../app');
 const	dbConnect = require('../utils/dbConnect');
 const	Blog = require('../models/Blog');
+const	User = require('../models/User');
 const	helper = require('./test_helper');
+const	userHelper = require('./users_test_helper');
 
 const	api = supertest(app);
 const	url = '/api/blogs';
+const	userUrl = '/api/users';
+const	loginUrl = '/api/login';
+let		token;
 
 beforeEach(async() =>
 {
+	let	users;
 	await Blog.deleteMany({});
+	await User.deleteMany({});
 
 	const	blogObjects = helper.initialBlogs.map(blog => new Blog(blog));
-	const	promiseArray = blogObjects.map(blog => blog.save());
+	let		promiseArray = blogObjects.map(blog => blog.save());
+	promiseArray = promiseArray.concat(userHelper.initialUsers.map(user => api.post(userUrl).send(user)));
 
 	await Promise.all(promiseArray);
-}, 10000)
+	users = await api.get(userUrl);
+	users.body[0].password = 'jake';
+	console.log(users.body[0], userHelper.initialUsers[0]);
+	let	login = await api.post(loginUrl).send(users.body[0]);
+	token = login.body.token;
+}, 100000)
 
 afterAll(() =>
 {
@@ -62,7 +75,7 @@ describe('GET request testing', () =>
 
 describe('POST requests testing', () =>
 {
-	test('POST request creates a new blog', async () =>
+	test('if no authorization token is provided an error is thrown', async () =>
 	{
 		//arrange
 		let	response;
@@ -73,7 +86,25 @@ describe('POST requests testing', () =>
 		response = await api.post(url).send(blog);
 		blogsAtEnd = await helper.blogsInDb();
 		//assert
-		expect(response.status).toBe(200);
+		expect(response.status).toBe(401);
+		expect(response.unauthorized).toBe(true);
+		expect(response.body.error).toBeDefined();
+		expect(response.body.error).toBe('invalid token');
+		expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length);
+	}, 100000)
+
+	test('POST request creates a new blog', async () =>
+	{
+		//arrange
+		let	response;
+		let	blog;
+		let	blogsAtEnd;
+		//act
+		blog = helper.blogObject;
+		response = await api.post(url).set('Authorization', `bearer ${token}`).send(blog);
+		blogsAtEnd = await helper.blogsInDb();
+		//assert
+		expect(response.status).toBe(201);
 		expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length + 1);
 	})
 
@@ -84,12 +115,12 @@ describe('POST requests testing', () =>
 		let	blog;
 		//act
 		blog = helper.likesMissing;
-		await api.post(url).send(blog);
+		await api.post(url).set('Authorization', `bearer ${token}`).send(blog);
 		response = await helper.blogsInDb();
 		//assert
 		expect(blog.likes).toBeUndefined();
 		expect(response[response.length - 1].likes).toBe(0);
-	})
+	}, 100000)
 
 	test('if title or url are missing, the backend throws an error', async () =>
 	{
@@ -98,20 +129,20 @@ describe('POST requests testing', () =>
 		let	blog;
 		//act
 		blog = helper.titleMissing;
-		response = await api.post(url).send(blog);
+		response = await api.post(url).set('Authorization', `bearer ${token}`).send(blog);
 		//assert
 		expect(response.status).toBe(400);
 		//act2
 		blog = helper.urlMissing;
-		response = await api.post(url).send(blog);
+		response = await api.post(url).set('Authorization', `bearer ${token}`).send(blog);
 		//assert2
 		expect(response.status).toBe(400);
-	})
+	}, 100000)
 })
 
 describe('DELETE requests testing', () =>
 {
-	test('DELETE request deletes specified blog', async () =>
+	test('DELETE w/o token throws error', async () =>
 	{
 		//arrange
 		let	blogs;
@@ -124,30 +155,32 @@ describe('DELETE requests testing', () =>
 		response = await api.delete(`${url}/${toDelete.id}`);
 		blogsAtEnd = await helper.blogsInDb();
 		//assert
-		expect(response.status).toBe(204);
-		expect(blogsAtEnd).not.toContainEqual(toDelete);
-		expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length - 1);
+		expect(response.status).toBe(401);
+		expect(response.body.error).toBe('invalid token');
+		expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length);
 	})
-})
-describe('DELETE requests testing', () =>
-{
-	test('DELETE request deletes specified blog', async () =>
+
+	test('DELETE request eliminates a blog entry', async () =>
 	{
 		//arrange
 		let	blogs;
 		let	toDelete;
 		let	response;
 		let	blogsAtEnd;
+		let	users;
 		//act
-		blogs = await helper.blogsInDb();
-		toDelete = blogs[0];
-		response = await api.delete(`${url}/${toDelete.id}`);
+		blogs = await api.get(url);
+		users = await api.get(userUrl);
+		blogs.body[0].user = users.body[0].id;
+		const updatedBlog = await api.put(`${url}/${blogs.body[0].id}`).send(blogs.body[0]);
+		toDelete = updatedBlog.body;
+		response = await api.delete(`${url}/${toDelete.id}`).set('Authorization', `bearer ${token}`);
 		blogsAtEnd = await helper.blogsInDb();
 		//assert
 		expect(response.status).toBe(204);
 		expect(blogsAtEnd).not.toContainEqual(toDelete);
 		expect(blogsAtEnd).toHaveLength(helper.initialBlogs.length - 1);
-	})
+	}, 100000)
 })
 
 describe('PUT requests testing', () =>
